requirements.txt
sklearn==0.0
scipy==1.0.0
msgpack-numpy==0.4.1
numpy==1.14.0
kenlm==0.0.0
CyHunspell==1.2.1
editdistance==0.4
spacy==2.0.5
